# resources/swagger.yaml
swagger: "2.0"
info:
  description: "Documentacion de la API Rest para la materia Aplicaciones Distribuidas de la Universidad Argentina de la Empresa (UADE), realizado en swagger/OpenAPI 3."
  version: "1.0.0"
  title: "Medical managment Backend"
  contact:
    name: "Yanzon, Carlos Santiago"
    email: "c.s.yanzon@gmail.com"
    url: "https://www.linkedin.com/in/carlos-santiago-yanzon/"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
...
schemes:
  - "http"
paths:
...
  /booking:
    post:
      tags:
        - "Turno"
      summary: "Crear un turno"
      description: "Crea un turno asociando un medico y un paciente."
      operationId: "addBooking"
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Posee la informacion necesaria para poder crear un turno."
          required: true
          schema:
            $ref: "#/definitions/Turno"
      responses:
        "200":
          description: "Se crea el turno con exito."
        "300":
          description: "El turno se intnteo crear fuera de los dos meses"
        "400":
          description: "Datos incorrectos."
    delete:
      tags:
        - "Turno"
      summary: "Cancelar un turno por centro medico"
      description: "Cancela un turno asociando de un paciente por parte del centro medico"
      operationId: "cancelBookingByMedicCentre"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "body"
          in: "body"
          description: "Id del turno."
          required: true
          schema:
            type: "object"
            properties:
              bookingId:
                type: "integer"
                example: "1"
      responses:
        "200":
          description: "Se cancela el turno con exito"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Se ha cancelado el turno con exito"
        "400":
          description: "Error al cancelrar el turno"
  /booking/confirmBooking:
    put:
      tags:
        - "Turno"
      summary: "Confirmar un turno"
      description: "Confirma un turno asociando de un paciente siempre que este se encuentre 12 hs antes."
      operationId: "confirmBooking"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "body"
          in: "body"
          description: "Id del turno."
          required: true
          schema:
            type: "object"
            properties:
              bookingId:
                type: "integer"
                example: "1"
      responses:
        "200":
          description: "Se confirma el turno con exito"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Se ha el turno confirmado con exito"
        "300":
          description: "Se intenta confirmar el turno antes"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "No se puede confirmar un turno una hora antes"
  /booking/cancelBooking:
    put:
      tags:
        - "Turno"
      summary: "Cancelar un turno"
      description: "Cancela un turno asociando de un paciente siempre que este se encuentre 2 meses proximos a la fecha."
      operationId: "cancelBooking"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "body"
          in: "body"
          description: "Id del turno."
          required: true
          schema:
            type: "object"
            properties:
              bookingId:
                type: "integer"
                example: "1"
      responses:
        "200":
          description: "Se cancela el turno con exito"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Se ha cancelado el turno con exito"
        "300":
          description: "Devuelve cuando se desea cancelar el turno 12 horas antes generara cargas adicionales"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Cancelar el turno 12 horas antes generara cargas adicionales"
        "301":
          description: "El turno ha expirado (se desea cancelar un turno viejo)"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "El turno ha expirado"
        "400":
          description: "Error al cancelar el turno"
...
definitions:
...
  Turno:
    type: "object"
    required:
      - "patientId"
      - "medicId"
      - "specialityId"
      - "day"
      - "time_start"
      - "time_end"
    properties:
      bookingId:
        type: "integer"
        format: "int64"
      day:
        type: "string"
        example: "2020-02-01"
      # category:
      # $ref: "#/definitions/Category"
      status:
        type: "string"
        example: "confirmado"
      time_start:
        type: "string"
        example: "10:30"
      time_end:
        type: "string"
        example: "11:00"
      patientId:
        type: "string"
        example: "p-1"
      medicId:
        type: "string"
        example: "m-1"
      specialityId:
        type: "integer"
        example: "1"
...